refactor(tempo): limit session PR to opts plumbing
