+++ crates/cast/src/cmd/batch_mktx.rs
@@ -6,9 +6,7 @@
 use crate::{
     call_spec::CallSpec,
     tempo,
-    tx::{
-        self, CastTxBuilder, ensure_session_compatible_raw_modes, resolve_wallet_or_session_signer,
-    },
+    tx::{self, CastTxBuilder},
 };
 use alloy_consensus::SignableTransaction;
 use alloy_eips::eip2718::Encodable2718;
@@ -70,20 +68,16 @@ impl BatchMakeTxArgs {
         // Resolve `--tempo.lane <name>` against the lanes file (default
         // `<root>/tempo.lanes.toml`) and populate `tx.tempo.nonce_key` from the lane.
         let resolved_lane = resolve_lane(&mut tx.tempo, &config.root)?;
-        let chain = utils::get_chain(config.chain, &provider).await?;
 
-        let resolved_signer =
-            resolve_wallet_or_session_signer(&tx.tempo, &eth.wallet, chain.id()).await?;
-        if resolved_signer.is_session {
-            ensure_session_compatible_raw_modes(raw_unsigned, ethsign)?;
-        }
-        let (signer, tempo_access_key) = resolved_signer.into_parts();
+        // Resolve signer to detect keychain mode
+        let (signer, tempo_access_key) = eth.wallet.maybe_signer().await?;
 
         // Parse all call specs
         let call_specs: Vec<CallSpec> =
             calls.iter().map(|s| CallSpec::parse(s)).collect::<Result<Vec<_>>>()?;
 
         // Get chain for parsing function args
+        let chain = utils::get_chain(config.chain, &provider).await?;
         let etherscan_config = config.get_etherscan_config_with_chain(Some(chain)).ok().flatten();
         let etherscan_api_key = etherscan_config.as_ref().map(|c| c.key.clone());
         let etherscan_api_url = etherscan_config.map(|c| c.api_url);
